Adopt ConfigurationDoneRequestHandler
